// web-angular-nodejs/src/app/core/models/cuttOff.model.ts
export class CuttOffModel {
  id: number = 0;
  areaOfInterestId: number = 0;
  areaOfInterestName: string = '';
  driveId: number = 0;
  cutoffMarks: number = 0;
  autoCutOffMarks: number = 0;
  fromAddDrive: boolean = false;
  paperId: number = 0;
  paperName: string = '';
  totalMarks: number = 0;
  driveName: string = '';
  constructor() {}
}